refactor(design): extract rasterization helpers
